refactor(admin): split routes.py into routes/ subpackage
Replace the 1,489-line admin/routes.py monolith with a focused
subpackage of six modules grouped by concern:

  routes/_shared.py      — shared helpers (_qp, _mask_api_key, etc.)
  routes/auth.py         — HTML serving, login, rate limiting
  routes/config.py       — config CRUD + fetch/bulk-add models
  routes/keys.py         — gateway API key management
  routes/observability.py — metrics, request log, network diagnostics
  routes/testing.py      — async model test tasks

routes/__init__.py re-exports everything and houses register_admin_routes(),
so no changes are needed in app.py or any other caller.
