+++ src/llm_rosetta/gateway/admin/routes/_shared.py
@@ -1,0 +1,119 @@
+"""Shared helpers used by multiple admin route modules."""
+
+from __future__ import annotations
+
+import re
+from typing import Any, overload
+
+from llm_rosetta._vendor.httpserver import JSONResponse, Response
+
+from ...config import GatewayConfig, load_config, write_config  # noqa: F401
+
+_ENV_VAR_RE = re.compile(r"^\$\{.+\}$")
+
+
+@overload
+def _qp(request: Any, key: str) -> str | None: ...
+
+
+@overload
+def _qp(request: Any, key: str, default: str) -> str: ...
+
+
+def _qp(request: Any, key: str, default: str | None = None) -> str | None:
+    """Extract a single query param value (httpserver convenience)."""
+    vals = request.query_params.get(key)
+    if vals:
+        return vals[0]
+    return default
+
+
+def _mask_api_key(value: str) -> str:
+    """Mask a literal API key, leaving env-var placeholders intact."""
+    if _ENV_VAR_RE.match(value):
+        return value
+    if len(value) <= 8:
+        return "***"
+    return value[:4] + "***" + value[-4:]
+
+
+def _get_config_path(request: Any) -> str | None:
+    """Return the config file path stored on the app object."""
+    return getattr(request.app, "config_path", None)
+
+
+def _reload_gateway_config(request: Any, config_path: str) -> GatewayConfig:
+    """Re-read config from disk, rebuild GatewayConfig, swap into app state."""
+    import llm_rosetta.gateway.app as _app_mod
+
+    raw = load_config(config_path)
+    new_config = GatewayConfig(raw)
+    _app_mod._config = new_config
+    request.app.gateway_config = new_config
+
+    _sync_auth_middleware(request.app, new_config)
+
+    return new_config
+
+
+def _sync_auth_middleware(app: Any, config: GatewayConfig) -> None:
+    """Update the auth hook's state for hot-reload."""
+    auth_state = getattr(app, "auth_state", None)
+    if auth_state is not None:
+        auth_state.key_set = config.api_key_set
+        auth_state.labels = dict(config.api_key_labels)
+
+
+def _build_provider_entry(
+    body: dict[str, Any],
+    api_key: str,
+    base_url: str,
+    existing_providers: dict[str, Any],
+    resolve_name: str,
+) -> dict[str, Any]:
+    """Build a provider entry dict from request body, resolving masked keys."""
+    if "***" in api_key and resolve_name in existing_providers:
+        api_key = existing_providers[resolve_name].get("api_key", api_key)
+
+    entry: dict[str, Any] = {"api_key": api_key, "base_url": base_url}
+
+    provider_type = body.get("type")
+    if provider_type:
+        entry["type"] = provider_type
+
+    if "proxy" in body:
+        proxy = body["proxy"]
+        if proxy:
+            entry["proxy"] = proxy
+
+    if resolve_name in existing_providers:
+        existing_enabled = existing_providers[resolve_name].get("enabled")
+        if existing_enabled is not None:
+            entry["enabled"] = existing_enabled
+
+    return entry
+
+
+def _handle_provider_rename(
+    data: dict[str, Any], rename_from: str, name: str
+) -> Response | None:
+    """Handle provider rename: remove old entry, update model refs."""
+    providers = data.get("providers", {})
+    if rename_from not in providers:
+        return JSONResponse(
+            {"error": f"Original provider '{rename_from}' not found"},
+            status_code=404,
+        )
+    if name in providers:
+        return JSONResponse(
+            {"error": f"Provider '{name}' already exists"},
+            status_code=409,
+        )
+    del providers[rename_from]
+    models = data.get("models", {})
+    for model_name, model_val in models.items():
+        if isinstance(model_val, str) and model_val == rename_from:
+            models[model_name] = name
+        elif isinstance(model_val, dict) and model_val.get("provider") == rename_from:
+            model_val["provider"] = name
+    return None
